[Refactor] #1701 source candidate evidence 수집 공통 로직 모듈 추출 — 수집기 중복 제거
KRIC/data.go.kr evidence 수집기 두 파일이 XML 안전 파서·sanitize·해시·수집
파이프라인을 거의 그대로 복제하고 있어 SonarCloud 신규 코드 중복도가
임계를 초과했다. 공용 로직을 tools/datapack/lib/source-candidate-evidence-collector.mjs로
추출하고 두 수집기는 서비스별 origin 검증·env 변수명·산출물 접두사만
옵션으로 주입해 위임하도록 재구성했다. 외부 계약(export 이름, CLI 인자
형식, env 변수명, 산출물 파일명·필드 순서, 에러 메시지)은 변경하지 않았다.
